Create injective function
In this example, the isInjective function takes an array arr and its size n as arguments. It checks if each element in the array is unique (i.e., if the function is injective) by using an unordered_set to keep track of seen values. If any element is found more than once, the function returns false.

In the main function, we create an example array arr and compute its size n. We then call the isInjective function with these arguments and print whether the function is injective or not based on its return value.

Note that an injective function is also known as a one-to-one function or a "no duplicates" function.